pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openapi2cli"
version = "0.1.0"
description = "Generate CLI tools from OpenAPI specs — built for AI agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Olaf", email = "olaf.bot@agentmail.to"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Internet :: WWW/HTTP",
]
keywords = ["openapi", "cli", "code-generation", "api", "swagger", "agents"]
dependencies = [
    "click>=8.0",
    "requests>=2.25",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
openapi2cli = "openapi2cli.cli:main"

[project.urls]
Homepage = "https://github.com/Olafs-World/openapi2cli"
Repository = "https://github.com/Olafs-World/openapi2cli"
Documentation = "https://github.com/Olafs-World/openapi2cli#readme"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (may hit real APIs)",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
